FeNiBilayer: mv auxiliary classes to implementation
